# Conflicts:
#	scripts/run_tutorial_transformations.sh

+++ scripts/run_local_notebook.sh
@@ -1,7 +1,12 @@
-#!/usr/bin/env bash
+#!/bin/bash
 
 set -e
 
-ROOT=$(pwd)/../
+if [ ! -d "blurr/core" ]; then
+    echo "test.sh should be run from blurr's base directory."
+    exit 1
+fi
+
+ROOT=$(pwd)
 
 docker run -it --rm -p 8888:8888  -v ${ROOT}:/home/jovyan jupyter/datascience-notebook
